jai-module-hacks.patch: Update for beta 0.1.096, generate POSIX and Socket bindings
The x64 bindings for those did NOT work, at least not with the Raspberry
Pi toolchain I tested with. This was causing havok whenever you used
POSIX too much.

The Socket ones may not be necessary, but since it was different at all,
I decided to include them just in case.