jai-module-hacks.patch: Fix Simp shader, late-load GL, GLX, X11
I'm basically extending what `GL.jai` was already doing, but now all of
these libraries are entirely loaded using `dlopen` and `dlsym`/
`glXGetProcAddress`. This is so that none of them show up as dynamic
link dependencies (as seen through, for example, `readelf -d`). This
makes it easy to cross-compile full executables - a full target sysroot
now need not be present. And, failure to load these libraries now can be
gracefully responded to and not prevent program startup.